// frontend/src/app/api/model/proizvodDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProizvodDTO { 
    idProizvod?: number;
    opis?: string;
    nikotinSadrzaj?: number;
    idTipProizvoda?: number;
}