// frontend/src/app/api/model/batchUnosNikotinaDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProizvodUnosDTO } from './proizvodUnosDTO';


export interface BatchUnosNikotinaDTO { 
    idKorisnik?: number;
    datum?: string;
    proizvodi?: Array<ProizvodUnosDTO>;
}